modify topHot by gst

+++ content-api/src/main/java/com/me2me/content/mapper/ContentMapper.java
@@ -142,7 +142,7 @@ public interface ContentMapper {
 
     List<Content> loadHottestTopsContent(@Param("flag") int flag);
 
-    List<Content> loadNewestContent(@Param("sinceId") int sinceId, @Param("flag") int flag);
+    List<Content> loadNewestContent(@Param("sinceId") long sinceId, @Param("flag") int flag);
 
     int countFragment(CountFragmentDto countFragmentDto);
